requirements.txt
openai-agents
markdown
streamlit
python-dotenv
mcp